[PATCH v2 03/11] tests/qtest/boot-serial-test: Build command line  using GString API

+++ m
@@ -2,57 +2,57 @@ Return-Path: <qemu-devel-bounces+lists+qemu-devel=lfdr.de@nongnu.org>
 X-Original-To: lists+qemu-devel@lfdr.de
 Delivered-To: lists+qemu-devel@lfdr.de
 Received: from lists.gnu.org (lists.gnu.org [209.51.188.17])
-	by mail.lfdr.de (Postfix) with ESMTPS id 60448673CE5
-	for <lists+qemu-devel@lfdr.de>; Thu, 19 Jan 2023 15:59:32 +0100 (CET)
+	by mail.lfdr.de (Postfix) with ESMTPS id 061BF673CF6
+	for <lists+qemu-devel@lfdr.de>; Thu, 19 Jan 2023 16:01:52 +0100 (CET)
 Received: from localhost ([::1] helo=lists1p.gnu.org)
 	by lists.gnu.org with esmtp (Exim 4.90_1)
 	(envelope-from <qemu-devel-bounces@nongnu.org>)
-	id 1pIWNS-00066U-U0; Thu, 19 Jan 2023 09:58:54 -0500
+	id 1pIWNX-0006CE-Oa; Thu, 19 Jan 2023 09:58:59 -0500
 Received: from eggs.gnu.org ([2001:470:142:3::10])
  by lists.gnu.org with esmtps (TLS1.2:ECDHE_RSA_AES_256_GCM_SHA384:256)
- (Exim 4.90_1) (envelope-from <philmd@linaro.org>) id 1pIWNR-00065I-7z
- for qemu-devel@nongnu.org; Thu, 19 Jan 2023 09:58:53 -0500
-Received: from mail-wm1-x32d.google.com ([2a00:1450:4864:20::32d])
+ (Exim 4.90_1) (envelope-from <philmd@linaro.org>) id 1pIWNW-0006BK-Kx
+ for qemu-devel@nongnu.org; Thu, 19 Jan 2023 09:58:58 -0500
+Received: from mail-wm1-x32e.google.com ([2a00:1450:4864:20::32e])
  by eggs.gnu.org with esmtps (TLS1.2:ECDHE_RSA_AES_128_GCM_SHA256:128)
- (Exim 4.90_1) (envelope-from <philmd@linaro.org>) id 1pIWNP-0003Xh-Ox
- for qemu-devel@nongnu.org; Thu, 19 Jan 2023 09:58:52 -0500
-Received: by mail-wm1-x32d.google.com with SMTP id
- o17-20020a05600c511100b003db021ef437so1455828wms.4
- for <qemu-devel@nongnu.org>; Thu, 19 Jan 2023 06:58:51 -0800 (PST)
+ (Exim 4.90_1) (envelope-from <philmd@linaro.org>) id 1pIWNV-0003Y7-1x
+ for qemu-devel@nongnu.org; Thu, 19 Jan 2023 09:58:58 -0500
+Received: by mail-wm1-x32e.google.com with SMTP id
+ iv8-20020a05600c548800b003db04a0a46bso3152905wmb.0
+ for <qemu-devel@nongnu.org>; Thu, 19 Jan 2023 06:58:56 -0800 (PST)
 DKIM-Signature: v=1; a=rsa-sha256; c=relaxed/relaxed; d=linaro.org; s=google;
  h=content-transfer-encoding:mime-version:references:in-reply-to
  :message-id:date:subject:cc:to:from:from:to:cc:subject:date
  :message-id:reply-to;
- bh=Dbql9l7vSKPentsKxfsPytXlVVV9PrWNWaeeTqrtx+A=;
- b=oEOk9AIQ/Fvj26tENwefQ4Gk1I/NGhdiZwSqLMwxte9ZzMTIZAKIgE0pXqFONYj2gq
- FBw0altoZ/+Dc6qpuF3XQFDmPdqPl1EtLggluXdCwy59w6lp9a9fmTtfLYAAkDFz+p0x
- +8PGA0AXy/NPs4dsh9ZhAO7Moqmt31MLqkafg0QNMGpPwcu3GwF5oZjUb4ajojO1plEo
- U17DUwgu1GN0dU1FgDzHiMXWkCQPUdE18u/gjClQxzFAv0AUD8y7Q2gEaB0/KNMlucCv
- nOVmyeyTO0SSKJ4OocqNVWHkjCkMlLKpqmV1Q+XYEPIXiP2FKSZPz1xG9ispuXd2BPxU
- 5gMg==
+ bh=es3Dw8MLiqLXATzoVsMpPJx/0iri3aBMV7qIVLaAL/4=;
+ b=s0S64k3YCGevc0R0oTZIH8fIMdDLMa6BHjCYopC0QoEYw3dMbRGT2SJkPn2oxAS29S
+ 5pgEK/hzdaM3N0gsH93FntPpHWRMX82EeqE0lCxAI2SPdHDeNVc58iPG0ZJGtvYelZop
+ 1BEuzraGBDQXNsM7Sw8QBMHAt/Edu+h1OKQ+0opkndkrNh9FEhRhJxq0zG9GZ0i9bJN0
+ aXrxlghJiCzsjgRBeHAsDkyaGL7kHdiOmxUeVcENUvBEi78eTe4mo2Auw2gRm/0YBuqN
+ M3G8d4NO7r6PG8vZU8QU8TGcGc52y0AlU/Ttjm4Xf92iyeI/wYYMlHcDYNTtlf1aEBRs
+ jFow==
 X-Google-DKIM-Signature: v=1; a=rsa-sha256; c=relaxed/relaxed;
  d=1e100.net; s=20210112;
  h=content-transfer-encoding:mime-version:references:in-reply-to
  :message-id:date:subject:cc:to:from:x-gm-message-state:from:to:cc
  :subject:date:message-id:reply-to;
- bh=Dbql9l7vSKPentsKxfsPytXlVVV9PrWNWaeeTqrtx+A=;
- b=dQVbOALwtwO0Zu648+TuWDNrMkyLAOY4aezO2PiKlu5m+HgEAnUh8MGzDs9KF9wLL+
- wZRPFXrQD2goG9Nh5TQrkSK9L6vema+wmS9grOmTDma3P4pEeHN2Jhc4v1hbF4PPd2un
- u5ZSP3ATuSeHsXosqSfcjS1MXuEO47OhhE0B5FIJTZDSEv/jLSpR9vVQnrHkbxax3pob
- a3j1a2mIJOP13Alb2IbCwlF+hc2vyLAAbYU3cOkXTBHME+UMXM0X+IpFxjV36Qrq/Hzx
- inqZMI6i14zA7H6aT5ShZS4OsG+YZDakAsa5eff1fuw1z/CpHfW9CPXc3xuNVnnDqFNh
- dxcg==
-X-Gm-Message-State: AFqh2krMzgzQQwuDwh031ogRPGBH7c4jz89wGLJKoub/JE6nTNLnNM6H
- 2f04ELd5u4KflKxpBWljatz7CjEen6PHvd4f
-X-Google-Smtp-Source: AMrXdXuOQI84rpnI1QJS6IiRmJ5z3yxqzf+hKCeqk2kPRceP5++f+U+G0g6BLlI8yZpkpPR3vfhNsw==
-X-Received: by 2002:a05:600c:4b0f:b0:3d6:2952:679b with SMTP id
- i15-20020a05600c4b0f00b003d62952679bmr10446693wmp.34.1674140330223; 
- Thu, 19 Jan 2023 06:58:50 -0800 (PST)
+ bh=es3Dw8MLiqLXATzoVsMpPJx/0iri3aBMV7qIVLaAL/4=;
+ b=5rtbWy6sVNNRNMjhFiOePv/Tns3D1xjDaVVxA8wTIycC4h3H5Ve+uncRB2D2AY0Dp6
+ qM0bnX3PGXqg2qOs5uTswMqSqam/fscxDYqlc6FQHZMRvZ6wAxSVV73/qqVxZ5Wb3mft
+ wlDyTeKZY0DBCvwAlJEU2rWkG4PVRAhOGEt81b9HxpKFA6wx3PRipehPN0LA7izCHTIH
+ /7PnVkY9r1GuAqI9g/hbUY1RNzGEXhcpjzjpGOkWZarb6x9LUJQUryhB1n2ZAmD6+WFc
+ MbQyVeWmhZkGZ1dWVf8VOuyePGLEK2oot6jpBTBTQ1cqEbAJUHCK0DTo+vpHMVpMmfmM
+ sj6w==
+X-Gm-Message-State: AFqh2krmgHqb9T44hi5k+TmrpoXooX63DTqOXoddY5mU/5tju1QANIqW
+ 8ibRWQYp3qq+bkdmS0BfJBKaA/u5V+kufL3I
+X-Google-Smtp-Source: AMrXdXt/xVdtHaHj611XoLi6fwLUx5kTFQwwrwE2Mdj6MfzU9yOl+cf8V0AqkYuMe2uhzDpt0/IL3Q==
+X-Received: by 2002:a05:600c:198e:b0:3db:1d7e:c429 with SMTP id
+ t14-20020a05600c198e00b003db1d7ec429mr3422988wmq.40.1674140335259; 
+ Thu, 19 Jan 2023 06:58:55 -0800 (PST)
 Received: from localhost.localdomain ([81.0.6.76])
  by smtp.gmail.com with ESMTPSA id
- ay13-20020a05600c1e0d00b003cf71b1f66csm5521830wmb.0.2023.01.19.06.58.49
+ r1-20020a05600c35c100b003db06493ee7sm5425204wmq.47.2023.01.19.06.58.54
  (version=TLS1_3 cipher=TLS_CHACHA20_POLY1305_SHA256 bits=256/256);
- Thu, 19 Jan 2023 06:58:49 -0800 (PST)
+ Thu, 19 Jan 2023 06:58:54 -0800 (PST)
 From: =?UTF-8?q?Philippe=20Mathieu-Daud=C3=A9?= <philmd@linaro.org>
 To: qemu-devel@nongnu.org
 Cc: Juan Quintela <quintela@redhat.com>, Laurent Vivier <lvivier@redhat.com>,
@@ -60,18 +60,18 @@ Cc: Juan Quintela <quintela@redhat.com>, Laurent Vivier <lvivier@redhat.com>,
  "Dr. David Alan Gilbert" <dgilbert@redhat.com>, qemu-arm@nongnu.org,
  Fabiano Rosas <farosas@suse.de>, Thomas Huth <thuth@redhat.com>,
  =?UTF-8?q?Philippe=20Mathieu-Daud=C3=A9?= <philmd@linaro.org>
-Subject: [PATCH v2 02/11] tests/qtest/boot-serial-test: Simplify
- test_machine() a bit
-Date: Thu, 19 Jan 2023 15:58:29 +0100
-Message-Id: <20230119145838.41835-3-philmd@linaro.org>
+Subject: [PATCH v2 03/11] tests/qtest/boot-serial-test: Build command line
+ using GString API
+Date: Thu, 19 Jan 2023 15:58:30 +0100
+Message-Id: <20230119145838.41835-4-philmd@linaro.org>
 X-Mailer: git-send-email 2.38.1
 In-Reply-To: <20230119145838.41835-1-philmd@linaro.org>
 References: <20230119145838.41835-1-philmd@linaro.org>
 MIME-Version: 1.0
 Content-Type: text/plain; charset=UTF-8
 Content-Transfer-Encoding: 8bit
-Received-SPF: pass client-ip=2a00:1450:4864:20::32d;
- envelope-from=philmd@linaro.org; helo=mail-wm1-x32d.google.com
+Received-SPF: pass client-ip=2a00:1450:4864:20::32e;
+ envelope-from=philmd@linaro.org; helo=mail-wm1-x32e.google.com
 X-Spam_score_int: -20
 X-Spam_score: -2.1
 X-Spam_bar: --
@@ -94,67 +94,65 @@ List-Subscribe: <https://lists.nongnu.org/mailman/listinfo/qemu-devel>,
 Errors-To: qemu-devel-bounces+lists+qemu-devel=lfdr.de@nongnu.org
 Sender: qemu-devel-bounces+lists+qemu-devel=lfdr.de@nongnu.org
 
-Slighly modify test_machine() to simplify next commit review.
-
 Signed-off-by: Philippe Mathieu-Daudé <philmd@linaro.org>
 ---
- tests/qtest/boot-serial-test.c | 21 +++++++--------------
- 1 file changed, 7 insertions(+), 14 deletions(-)
+ tests/qtest/boot-serial-test.c | 23 +++++++++++++++--------
+ 1 file changed, 15 insertions(+), 8 deletions(-)
 
 diff --git a/tests/qtest/boot-serial-test.c b/tests/qtest/boot-serial-test.c
-index 3aef3a97a9..3a854b0174 100644
+index 3a854b0174..92890b409d 100644
 --- a/tests/qtest/boot-serial-test.c
 +++ b/tests/qtest/boot-serial-test.c
-@@ -227,7 +227,6 @@ static void test_machine(const void *data)
+@@ -226,14 +226,17 @@ static void test_machine(const void *data)
+     const testdef_t *test = data;
      g_autofree char *serialtmp = NULL;
      g_autofree char *codetmp = NULL;
-     const char *codeparam = "";
--    const uint8_t *code = NULL;
+-    const char *codeparam = "";
      QTestState *qts;
      int ser_fd;
++    g_autoptr(GString) cmd = g_string_new("");
  
-@@ -235,21 +234,13 @@ static void test_machine(const void *data)
+     ser_fd = g_file_open_tmp("qtest-boot-serial-sXXXXXX", &serialtmp, NULL);
      g_assert(ser_fd != -1);
      close(ser_fd);
  
--    if (test->kernel) {
--        code = test->kernel;
--        codeparam = "-kernel";
--    } else if (test->bios) {
--        code = test->bios;
--        codeparam = "-bios";
--    }
--
--    if (code) {
-+    if (test->kernel || test->bios) {
++    g_string_append_printf(cmd, "-M %s ", test->machine);
++    g_string_append(cmd, "-no-shutdown ");
++
+     if (test->kernel || test->bios) {
          ssize_t wlen;
          int code_fd;
- 
-         code_fd = g_file_open_tmp("qtest-boot-serial-cXXXXXX", &codetmp, NULL);
-         g_assert(code_fd != -1);
--        wlen = write(code_fd, code, test->codesize);
-+        wlen = write(code_fd, test->kernel ? : test->bios, test->codesize);
+@@ -243,19 +246,23 @@ static void test_machine(const void *data)
+         wlen = write(code_fd, test->kernel ? : test->bios, test->codesize);
          g_assert(wlen == test->codesize);
          close(code_fd);
++        g_string_append_printf(cmd, "%s %s ",
++                               test->kernel ? "-kernel " : "-bios ", codetmp);
      }
-@@ -258,12 +249,14 @@ static void test_machine(const void *data)
+ 
++    g_string_append_printf(cmd, "-chardev file,id=serial0,path=%s "
++                                "-serial chardev:serial0 ", serialtmp);
++
+     /*
       * Make sure that this test uses tcg if available: It is used as a
       * fast-enough smoketest for that.
       */
--    qts = qtest_initf("%s %s -M %s -no-shutdown "
-+    qts = qtest_initf("%s %s %s -M %s -no-shutdown "
-                       "-chardev file,id=serial0,path=%s "
-                       "-serial chardev:serial0 -accel tcg -accel kvm %s",
--                      codeparam, code ? codetmp : "", test->machine,
-+                      codeparam,
-+                      test->kernel ? "-kernel " : test->bios ? "-bios " : "",
-+                      codetmp ? : "", test->machine,
-                       serialtmp, test->extra);
--    if (code) {
-+    if (codetmp) {
+-    qts = qtest_initf("%s %s %s -M %s -no-shutdown "
+-                      "-chardev file,id=serial0,path=%s "
+-                      "-serial chardev:serial0 -accel tcg -accel kvm %s",
+-                      codeparam,
+-                      test->kernel ? "-kernel " : test->bios ? "-bios " : "",
+-                      codetmp ? : "", test->machine,
+-                      serialtmp, test->extra);
++    g_string_append(cmd, "-accel tcg ");
++    g_string_append(cmd, "-accel kvm ");
++    g_string_append(cmd, test->extra);
++
++    qts = qtest_init(cmd->str);
++
+     if (codetmp) {
          unlink(codetmp);
      }
- 
 -- 
 2.38.1
